Fix invalid pointer access
Fixed a couple places where we try to use an invalid pointer when
certain dependencies are missing, such as if the classd daemon is
not available or the GeoIP database can't be loaded.

+++ plugins/classify/classify.go
@@ -101,6 +101,11 @@ func PluginShutdown() {
 	daemonConnection = nil
 
 	// terminate the classd daemon
+	if daemonProcess.Process == nil {
+		logger.Err("The classd daemon process was invalid\n")
+		return
+	}
+
 	err := daemonProcess.Process.Kill()
 	if err != nil {
 		logger.Err("Error stopping classd daemon: %v\n", err)
